Still working on translation of multiobject scene (works for objects with 1 triangle

+++ include/matrix4x4.h
@@ -25,7 +25,7 @@ class Matrix4x4
     protected:
 
     private:
-        int multiplicate_row_col(Matrix4x4* other, int row, int column);
+        float multiplicate_row_col(Matrix4x4* other, int row, int column);
 
 };
